# openapi.yaml
openapi: 3.0.1
info:
  title: Graip.AI Rest API
  description: To get API key, go to Flow -> Settings -> Integration -> New Access Token. Additional information can be found at [https://graip.atlassian.net/wiki/x/AYBvOg](https://graip.atlassian.net/wiki/x/AYBvOg)
  contact:
    email: support@graip.ai
  version: 1.0.0
servers:
  - url: https://app.graip.ai/api
...
paths:
...
  /v1/{flowId}/request/extract:
    post:
      tags:
        - request
      summary: Extract data from document
      description: Upload a document, process it to extract data, and return the extracted information within a single HTTP request. By default, neither the document nor the extracted data will be stored, even if processing fails. Set a long timeout (10 minutes) for this API call to accommodate the operation.
      parameters:
        - name: flowId
          in: path
          description: ID of flow
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: ['file']
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                pagesToProcess:
                  type: string
                  description: comma delimited list of pages with ranges, e.g 1,2-4,5,-1
                persistData:
                  type: boolean
                  default: false
                  description: Should persist data after extraction?
                extractionTimeout:
                  type: integer
                  minimum: 30
                  maximum: 3600
                  default: 600
                  description: Maximum time in seconds to wait for extraction to complete. Must be between 30 and 3600 seconds (30s – 60min). Defaults to 600 seconds (10 minutes).
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/RequestWithData'
        400:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                ERROR_UNSUPPORTED_FORMAT:
                  summary: 'ERROR_UNSUPPORTED_FORMAT'
                  value:
                    error: 'ERROR_UNSUPPORTED_FORMAT'
                    details: 'File mime type is not supported in current flow'
                ERROR_TOO_MANY_PAGES:
                  summary: 'ERROR_TOO_MANY_PAGES'
                  value:
                    error: 'ERROR_TOO_MANY_PAGES'
                    details: 'Document contains too many pages'
                ERROR_COURRUPTED_FILE:
                  summary: 'ERROR_COURRUPTED_FILE'
                  value:
                    error: 'ERROR_COURRUPTED_FILE'
                    details: 'Cannot read a file, file may be damaged, corrupted or have a password'
                ERROR_PAGES_LIMIT_REACHED:
                  summary: 'ERROR_PAGES_LIMIT_REACHED'
                  value:
                    error: 'ERROR_PAGES_LIMIT_REACHED'
                    details: 'Monthly document page upload limit has been reached'
                ERROR_FILE_TOO_SMALL:
                  summary: 'ERROR_FILE_TOO_SMALL'
                  value:
                    error: 'ERROR_FILE_TOO_SMALL'
                    details: 'Uploaded file is too small, check file content'
                ERROR_FILE_TOO_LARGE:
                  summary: 'ERROR_FILE_TOO_LARGE'
                  value:
                    error: 'ERROR_FILE_TOO_LARGE'
                    details: "The file size exceeds the allowed limit."
                ERROR_FLOW_IS_DISABLED:
                  summary: 'ERROR_FLOW_IS_DISABLED'
                  value:
                    error: 'ERROR_FLOW_IS_DISABLED'
                    details: 'Operation is not allowed for disabled flow'
                ERROR_MALFORMED_INPUT:
                  summary: 'ERROR_MALFORMED_INPUT'
                  value:
                    error: 'ERROR_MALFORMED_INPUT'
                    details: 'Request input is malformed'

        401:
          $ref: '#/responses/ErrUnauthorized'
        403:
          $ref: '#/responses/ErrForbidden'
        404:
          $ref: '#/responses/ErrNotFound'
        500:
          $ref: '#/responses/ErrServerError'
      security:
        - graip_auth:
            - read:requests
            - write:requests
...
responses:
    ErrServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'ERROR_SERVER_ERROR'
              details:
                type: string
                example: 'Internal server error'
    ErrForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'ERROR_FORBIDDEN'
              details:
                type: string
                example: 'No enough permissions to access requested resource or entity'
    ErrNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'ERROR_NOT_FOUND'
              details:
                type: string
                example: 'Requested resource or entity is not found'
    ErrUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'ERROR_UNAUTHORIZED'
              details:
                type: string
                example: 'Invalid auth credentials'
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
    Request:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: 'Document 1234'
        approval_status:
          type: string
          enum:
            - APPROVED
            - REJECTED
            - PENDING
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - DRAFT
            - PROCESSING
            - ERROR
            - PROCESSED
        status_info:
          oneOf:
            - type: object
              nullable: true
              example: null
            - $ref: '#/components/schemas/ErrServerError'
            - $ref: '#/components/schemas/ErrHookExecutionFailed'
            - $ref: '#/components/schemas/ErrThirdPartyServiceNotAvaliable'
            - $ref: '#/components/schemas/ErrInvalidFileStructure'
            - $ref: '#/components/schemas/ErrExtractionFailed'
            - $ref: '#/components/schemas/ErrModelProcessingTimeout'
            - $ref: '#/components/schemas/ErrInsufficientTextQuality'
            - $ref: '#/components/schemas/ErrUnsupportedLanguage'
            - $ref: '#/components/schemas/ErrMlInference'
...
    RequestWithData:
      allOf:
        - $ref: '#/components/schemas/Request'
        - type: object
          properties:
            data:
              type: object
              description: Extracted data
...
    ErrServerError:
      type: object
      properties:
        error:
          type: string
          example: 'ERROR_SERVER_ERROR'
        details:
          type: string
          example: 'Internal server error'
    ErrHookExecutionFailed:
      type: object
      properties:
        error:
          type: string
          example: 'ERROR_HOOK_EXECUTION_FAILED'
        details:
          type: string
          example: 'Hook PRE_REQUEST execution is failed'
    ErrThirdPartyServiceNotAvaliable:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_THIRD_PARTY_SERVICE_NOT_AVAILABLE"
        details:
          type: string
          example: "A required third-party service is currently unavailable. Please try again later."
    ErrInvalidFileStructure:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_INVALID_FILE_STRUCTURE"
        details:
          type: string
          example: "The file structure is invalid or unreadable by the extraction model."
    ErrExtractionFailed:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_EXTRACTION_FAILED"
        details:
          type: string
          example: "Data extraction failed due to an unexpected issue with the content."
    ErrModelProcessingTimeout:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_MODEL_PROCESSING_TIMEOUT"
        details:
          type: string
          example: "The ML model took too long to process the file."
    ErrInsufficientTextQuality:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_INSUFFICIENT_TEXT_QUALITY"
        details:
          type: string
          example: "The document's text quality is too low for accurate extraction."
    ErrUnsupportedLanguage:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_UNSUPPORTED_LANGUAGE"
        details:
          type: string
          example: "The detected language is not supported by the extraction model."
    ErrMlInference:
      type: object
      properties:
        error:
          type: string
          example: "ERROR_ML_INFERENCE_ERROR"
        details:
          type: string
          example: "Error during ml inference."
...
  securitySchemes:
    graip_auth:
      type: apiKey
      in: header
      name: Authorization